Fixed make_emscripten_preload_data.rb failing if data dir not present.

+++ scripts/make_emscripten_preload_data.rb
@@ -28,8 +28,8 @@ puts "glare_core: #{glare_core}"
 puts "substrata_dir: #{substrata_dir}"
 puts "-----------------------------------------"
 
-FileUtils.rm_r("data", :verbose=>true)
-FileUtils.mkdir_p("data", :verbose => true) if !File.exists?("data") # Make data dir if it doesn't exist already.
+FileUtils.rm_r("data", :verbose=>true) if File.exists?("data")
+FileUtils.mkdir_p("data", :verbose => true)
 
 FileUtils.cp_r(substrata_dir + "/resources", "data", :verbose=>true)
 FileUtils.rm_r("data/resources/materials", :verbose=>true)
